+++ packages/ako_optional/CHANGELOG.md
@@ -51,8 +51,8 @@ Going forward, this CHANGELOG only contains changes of the Ako Optional package.
 - AkoFutureOptional, a widget to simplify usage of FutureBuilder with Optionals
 - The `Optional` library is added as a dependency, and exported via `package:ako_core/optional.dart`
 
-[unreleased]: https://github.com/ATVG-Studios/Ako/compare/v0.9.0...HEAD
-[0.9.0]: https://github.com/ATVG-Studios/Ako/compare/v0.8.0...v0.9.0
+[unreleased]: https://github.com/ATVG-Studios/Ako/compare/ako_optional-v0.9.0...HEAD
+[0.9.0]: https://github.com/ATVG-Studios/Ako/compare/v0.8.0...ako_optional-v0.9.0
 [0.8.0]: https://github.com/ATVG-Studios/Ako/compare/v0.7.1...v0.8.0
 [0.5.0]: https://github.com/ATVG-Studios/Ako/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/ATVG-Studios/Ako/compare/v0.3.0...v0.4.0